+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+**2.4.8**
+* Fix Bug In ``extract_from_bl_instruction``
+* Update LIEF And GLFW
+* Allow Mods To Access The Original GLFW Keycode For Key Events (But Better)
+* More Accurate Sound
+
 **2.4.7**
 * Improve Server Performance
 * Add ``Add Biome Colors To Grass`` Feature Flag (Disabled By Default)